fix(store): ensure there are no cyclic dependencies mucking things up
